requirements.txt
python-dotenv==0.20.0
tweepy==4.10.0
